+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file. The format 
 
 <hr />
 
+## [2.1.52] - 2025-12-26
+Summary: Update.
+
+### Changed
+- Updated staged files.
+
+### Files
+- assets/css/analyser.css
+- assets/css/schema.css
+- assets/js/schema.js
+- includes/admin/schema-view-preview.php
+
+
 ## [2.1.51] - 2025-12-26
 Summary: Devnotes.